L2-993: CR changes
