+++ frontend/src/tests/lib/utils/sns.utils.spec.ts
@@ -4,12 +4,14 @@ import { SnsMetadataResponseEntries } from "@dfinity/sns";
 import {
   getSwapCanisterAccount,
   mapAndSortSnsQueryToSummaries,
+  mapOptionalToken,
 } from "../../../lib/utils/sns.utils";
 import { mockIdentity } from "../../mocks/auth.store.mock";
 import {
   mockDerived,
   mockQueryMetadata,
   mockQueryMetadataResponse,
+  mockQueryTokenResponse,
   mockSnsSummaryList,
   mockSummary,
   mockSwapInit,
@@ -276,6 +278,14 @@ describe("sns-utils", () => {
         controller: mockIdentity.getPrincipal(),
       });
       expect(expectedAccount).toBeInstanceOf(AccountIdentifier);
+    });
+  });
+
+  describe("mapOptionalToken", () => {
+    it("should return token", () => {
+      const token = mapOptionalToken(mockQueryTokenResponse);
+      expect(token?.name).toBeDefined();
+      expect(token?.symbol).toBeDefined();
     });
   });
 });
